Unexpected HTML escaping in macro concatenation
When macro A concatenates strings with the output of macro B, and both
macros mark their output as raw, the strings from macro A are
unexpectedly HTML escaped.  This only happens with the concatenation
operator when used inside a macro.  HTML escaping is not applied if the
join method is used instead of the concatenation operator, or if the
concatenation operator is used outside of a macro.